only enable alt-arch tools/clients when available
This patch adds a check when parsing the tool config files:
Clients for alternative archs (e.g. for cross-exec) injection
are only appended if the path to the client lib actually
exists.

This is a preparation for system-wide deployments, where this
check has to be performed at runtime as users might only
install the dr library for a single architecture.

Previously having invalid paths in the tool config file
worked if the respective client was not requested, but lead to a warning.
With this patch, the missing client is simply not registered at all.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>